+++ PusherMainView/PusherMainView/PusherViewController.swift
@@ -9,7 +9,7 @@ public final class PusherViewController: NSViewController {
     @IBOutlet private var appOrProjectIDTextField: NSTextField!
     @IBOutlet private var priorityTextField: NSTextField!
     @IBOutlet private var sandBoxCheckBox: NSButton!
-    @IBOutlet private var liveActivityCheckBox: NSButton!
+    @IBOutlet private var pushTypeTextField: NSTextField!
     @IBOutlet private var apnsCertificateRadioButton: NSButton!
     @IBOutlet private var apnsAuthTokenRadioButton: NSButton!
     @IBOutlet private var loadJSONFileButton: NSButton!
@@ -60,6 +60,7 @@ public final class PusherViewController: NSViewController {
         collapseIdTextField.placeholderString = "enter.apns.collapse.id".localized
         appOrProjectIDTextField.placeholderString = "enter.your.app.bundle.id".localized
         priorityTextField.placeholderString = "enter.apns.priority".localized
+        pushTypeTextField.placeholderString = "enter.apns.push.type".localized
         serverKeyTextField.placeholderString = "enter.fcm.server.key".localized
 
         priorityTextField.stringValue = "10"
@@ -163,7 +164,7 @@ public final class PusherViewController: NSViewController {
                 priority: priorityTextField?.integerValue ?? 10,
                 collapseID: collapseIdTextField.stringValue,
                 sandbox: sandBoxCheckBox.state.rawValue == 1,
-                liveActivity: liveActivityCheckBox.state.rawValue == 1))) { _ in })
+                pushType: pushTypeTextField.stringValue))) { _ in })
         }
     }
 
@@ -245,12 +246,12 @@ extension PusherViewController: NSTextViewDelegate {
     @IBOutlet private weak var priorityTextField: NSTextField!
     @IBOutlet private weak var collapseIdTextField: NSTextField!
     @IBOutlet private weak var sandBoxCheckBox: NSButton!
-    @IBOutlet private weak var liveActivityCheckBox: NSButton!
+    @IBOutlet private weak var pushTypeTextField: NSTextField!
     @IBOutlet private weak var serverKeyTextFieldContainter: NSView!
 
     private var iOSControls: [NSView] {
         [deviceTokenTextField, orLabel, selectDeviceButtonContainer, apnsButtonsContainer,
-         priorityTextField, collapseIdTextField, sandBoxCheckBox, liveActivityCheckBox]
+         priorityTextField, collapseIdTextField, sandBoxCheckBox, pushTypeTextField]
     }
     private var androidControls: [NSView] {
         [deviceTokenTextField, serverKeyTextFieldContainter, collapseIdTextField]