+++ PusherMainView/PusherMainView/String+Subscript.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension String {
+    subscript (range: Range<Int>) -> String {
+        let start = index(startIndex, offsetBy: range.lowerBound)
+        let end = index(startIndex, offsetBy: range.upperBound)
+        return String(self[start ..< end])
+    }
+}